RigisterViewController теперь AuthViewController и имеем один ViewController с логикой переиспользования в зависимости от действия пользователья (Аутентификация/регистрация). Добавлен ResetPasswordViewController.
Доработан UI AuthViewController и логика отображения UI элементов в зависимости от действия пользователя.
