+++ spravochnik_spz/Core/Entities/TypeViewController/AuthType.swift
@@ -1,0 +1,11 @@
+//
+//  AuthType.swift
+//  spravochnik_spz
+//
+//  Created by Swift Learning on 16.02.2023.
+//
+
+enum AuthType {
+    case auth
+    case register
+}
